QV partie Agenda

+++ src/DTO/EvenementEdt.php
@@ -3,7 +3,7 @@
 // @file /Users/davidannebicque/htdocs/intranetV3/src/DTO/EvenementEdt.php
 // @author davidannebicque
 // @project intranetV3
-// @lastUpdate 14/10/2020 09:06
+// @lastUpdate 14/10/2020 12:09
 
 namespace App\DTO;
 
@@ -12,7 +12,7 @@ class EvenementEdt
 {
     public ?string $date;
     public ?string $jour;
-    public ?int $duree;
+    public ?float $duree;
     public ?string $heure;
     public ?string $matiere;
     public ?string $personnel;
